Add a save button and a handle function
The function is just stubbed right now. But because Timeblock component mutates
the timeblock object, we can view what blocks are changed in the Day component
without needing to do any event propogation. Not sure if this is the right way
to do it but it works for now!

The next step is to take the blocks that have been checked and convert it to be
something that can be saved to the DB.