Write selected timeblocks to DB
This commit naively takes the start time as the first time block and the end
time as the last time block. Unfortunately, this means that gaps in selected
blocks are not taken into account. We're also implicitly relying on the
ordering. Which might not be obvious for someone new looking at the codebase.

I'm also not saving any task name with the selected time blocks. However, this
commit is enough to check that we can write to supabase through user
interaction!